% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getKEGGnames}
\alias{getKEGGnames}
\title{Get KEGG  names}
\usage{
getKEGGnames(x, verbose = TRUE)
}
\arguments{
\item{x}{a character vector of KEGG ids.}

\item{verbose}{verbose.}
}
\value{
A character vector with the corresponding KEGG names.
}
\description{
Finds the KEGG name form KEGG id.
}
\details{
Uses the library KEGG.db.

\code{x} may be a \code{data.frame}. In such case, GO ids are expected in its row names.
}
\examples{
getKEGGnames (c("00010", "00020", "BAD_KEGG"))
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\seealso{
\code{\link{getGOnames}}
}
\keyword{KEGG}
\keyword{names}

