% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOntology}
\alias{getOntology}
\title{Get GO term Ontology}
\usage{
getOntology(x, verbose = TRUE)
}
\arguments{
\item{x}{a character vector of GO ids.}

\item{verbose}{verbose.}
}
\value{
A character vector with the corresponding GO names.
}
\description{
Finds the ontology of a term from its id.
}
\details{
Uses the library GO.db.

\code{x} may be a \code{data.frame}.
In such case, GO ids are expected in its row names.
}
\examples{
getOntology (c("GO:0000018", "GO:0005788", "BAD_GO"))
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\seealso{
\code{\link{getGOnames}}, \code{\link{propagateGO}},
\code{\link{goLeaves}}, \code{\link{splitOntologies}},
\code{\link{getKEGGnames}}
}
\keyword{GO}
\keyword{names}
\keyword{ontology}

