% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mdGsa}
\alias{mdGsa}
\alias{multidimensionalGsa}
\alias{multivariateGsa}
\title{Multi-Dimensional Gene Set Analysis.}
\usage{
mdGsa(index, annot, p.adjust.method = "BY", family = quasibinomial(),
  verbose = TRUE, verbosity = 100, fulltable = FALSE,
  useColnames = TRUE, ...)
}
\arguments{
\item{index}{ranking index, generally a two column matrix.}

\item{annot}{an annotation list.}

\item{p.adjust.method}{p-value adjustment method for multiple testing.}

\item{family}{see \code{glm}.}

\item{verbose}{verbose.}

\item{verbosity}{integer indicating which iterations should be indicated if verbose = TRUE.}

\item{fulltable}{if TRUE, 'sd', 't' and 'convergence' indicator from the glm fit are included in the output.}

\item{useColnames}{if TRUE the names of the two first columns of the matrix 'index' are used in the results data.frame.}

\item{\dots}{further arguments to be pasted to \code{glm.fit}, for instance 'weights'.}
}
\value{
A \code{data.frame} with a row for each Gene Set or block.
Columns are:
\describe{
  \item{\code{N}:}{number of genes annotated to the Gene Set.}
  \item{\code{lor}:}{log Odds Ratio estimated for the Gene Set.}
  \item{\code{pval}:}{p-values associated to each log Odds Ratio.}
  \item{\code{padj}:}{adjusted p-values.}
  \item{\code{sd}:}{standard deviations associated to each log Odds Ratio.}
  \item{\code{t}:}{t statistic associated to each log Odds Ratio.}
}
Apart from the 'N' coefficient, all other indices appear in triplicate:
one coefficient for each genomic condition and a third one for the interaction.
}
\description{
Performs a Multi-Variate Gene Set Analysis for two genomic measurements.
}
\details{
'index' must be a numerical \code{matrix} or \code{data.frame} with at least two columns.
If there are more than three columns, the ranking indexes are taken form the two first one.
The remaining columns are used as covariates to correct for within the analysis.

Default p-value correction is "BY".

In the output data.frame there are three parameters of each type: 'lor', 'pval', \dots
one for each of the two genomic conditions analyzed and the third one for the interaction between them.
}
\examples{
\dontrun{
res <- mdGsa (rindexMat, annotList)
}
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\seealso{
\code{\link{uvGsa}}, \code{\link{mdPat}}, \code{glm.fit}, \code{p.adjust}
}
\keyword{GSA}
\keyword{gene}
\keyword{multidimensional}
\keyword{multivariate}
\keyword{set}

