% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mdPat}
\alias{mdPat}
\title{Multi-Dimensional Gene Set Analysis Pattern Classification.}
\usage{
mdPat(gsaout, cutoff = 0.05, pvalue = "padj")
}
\arguments{
\item{gsaout}{data.frame; output from mdGsa.}

\item{cutoff}{p-value cutoff for considering significant a Gene Set.}

\item{pvalue}{p-value column to be used. Default is named "padj" as in mdGsa output.}
}
\value{
A character vector indicating the pattern associated to each Gene Set.
}
\description{
Classifies significant patterns form a Multi-Variate Gene Set Analysis.
}
\details{
Sign of the three 'lor' and p-values are used to classify functional blocks.
The classification is done in the two dimensional space previously analyzed
by mdGsa.
}
\examples{
data (breast)
res <- mdGsa (ranking, annot)
cbind (res, mdPat (res))
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\references{
\href{fhttp://www.plosone.org/article/info\%3Adoi\%2F10.1371\%2Fjournal.pone.0010348}{Montaner et al. 2010 "Multidimensional Gene Set Analysis of Genomic Data." PLoS ONE.}
}
\seealso{
\code{\link{mdGsa}}, \code{\link{uvPat}}
}
\keyword{GSA}
\keyword{multidimensional}
\keyword{multivariate}
\keyword{pattern}

