% Generated by roxygen2 (4.0.2): do not edit by hand
\name{propagateGO}
\alias{propagateGO}
\title{Propagate Gene Ontology annotation.}
\usage{
propagateGO(annot, verbose = FALSE)
}
\arguments{
\item{annot}{annotation list or matrix.}

\item{verbose}{verbose}
}
\value{
A annotation matrix or list with the propagated annotation.
}
\description{
Genes annotated under a GO term inherit the annotation from
all its ancestors.
}
\details{
Uses the library GO.db.
}
\examples{
mat <- cbind (c("gene1", "gene2"), c("GO:0034390", "GO:0042889"))
mat
propagateGO (mat)

li <- list ('GO:0034390' = "gene1", 'GO:0042889' = "gene2")
li
propagateGO (li)
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\seealso{
\code{\link{annotMat2list}}, \code{\link{annotFilter}}
}
\keyword{GO}
\keyword{gene}
\keyword{ontology}
\keyword{propagate}

