% Generated by roxygen2 (4.0.2): do not edit by hand
\name{splitOntologies}
\alias{splitOntologies}
\title{Split an annotation list of GO terms by ontologies.}
\usage{
splitOntologies(annot, na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{annot}{annotation list.}

\item{na.rm}{if TRUE 'unknown' terms are excluded.}

\item{verbose}{verbose}
}
\value{
A list with tree components, one for each ontology.
A fourth component is included is some term could not be allocated to
any of the three GO ontologies.
}
\description{
Splits an annotation list of GO terms according to the ontology
to which each term belongs to.
}
\details{
Uses the information form the library GO.db.
If some id could not be associated to any ontology,
they are returned in an unknown ontology named "missing".
}
\examples{
getGOnames (c ("GO:0006915", "GO:0016020", "GO:0008152", "GO:0015288"))

annot <- list ("GO:0006915" = c ("g1"),
               "GO:0016020" = c ("g2", "g3"),
               "GO:0008152" = c ("g1", "g2", "g3"),
               "GO:0015288" = c ("g4", "g5"))
annot
splitOntologies (annot)
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\seealso{
\code{\link{propagateGO}}, \code{\link{goLeaves}}
}
\keyword{GO}
\keyword{ontology}
\keyword{split}

