% Generated by roxygen2 (4.0.2): do not edit by hand
\name{transferIndex}
\alias{transferIndex}
\title{Transfer a ranking index from regulatory elements,
such as miRNAs, to genes.}
\usage{
transferIndex(index, targets, method = "sum", verbose = TRUE,
  transferMatrix = FALSE)
}
\arguments{
\item{index}{ranking index, generally a numerical named vector.}

\item{targets}{a list describing the regulation blocks. See details.}

\item{method}{the method to collapse the gene information.
Currently 'sum' and 'average' are available.}

\item{verbose}{verbose.}

\item{transferMatrix}{if true the transference matrix is returned
instead of the ranking statistic.}
}
\value{
A ranking index transferred to the IDs in the elements of the
list \code{targets}.
}
\description{
Transfers the ranking index information from some regulatory elements
to their target genes.
It can be for instance used to infer gene regulation levels from miRNA
differential expression levels.
Afterwords, the inferred gene index can be explored in a
univariate gene set analysis using \code{uvGsa},
or in a multivariate gene set analysis using \code{mdGsa}.
}
\details{
The function was in principle designed to transfer a ranking index
defined for miRNAs to their target genes,
but it may also be used to deal with some other regulatory elements
as transcription factors for instance.

If we have for instance a t statistic accounting for miRNAs
differential expression levels,
we may add up (or average) all t values from the miRNAs that
regulate a gene in order to derive an interference score for that gene.
Thus we may get an interference index
which may be interpreted in terms of functional blocks or gene sets.

\code{index} may be a matrix.
In such case just the first column will be used.

\code{targets} should be a named list.
The names of the list have to coincide (or at least overlap)
with the names in the \code{index}.
The elements of the list are character vectors containing gene IDs.
In the miRNA example, each element of \code{targets}
will contain the genes regulated by a miRNA,
and the miRNA IDs will appear in the names of the \code{targets}
list as well as in the ranking index.
}
\examples{
## miRNA to gene list (targets)
targets <- list (mirna1 = "g1", mirna2 = c ("g1", "g2"), mirna3 = "g3")

## original index
index <- rnorm (5)
names (index) <- paste0 ("mirna", 1:5)

##transfered index
tindex <- transferIndex (index, targets)

##transformed (normalized) index
rindex <- indexTransform (tindex)

##NOTICE in this case:
index["mirna1"] + index["mirna2"] == tindex["g1"]

\dontrun{
res <- uvgsa (rindex, annot)
}
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\seealso{
\code{\link{uvGsa}}, \code{\link{mdGsa}},
\code{\link{indexTransform}}
}
\keyword{genes}
\keyword{index}
\keyword{miRNAs}
\keyword{to}
\keyword{transfer}

