% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GO_KEGG_names.R
\name{getGOnames}
\alias{getGOnames}
\title{Get Gene Ontology names}
\usage{
getGOnames(x, verbose = TRUE)
}
\arguments{
\item{x}{a character vector of GO ids.}

\item{verbose}{verbose.}
}
\value{
A character vector with the corresponding GO names.
}
\description{
Finds the GO name form GO id.
}
\details{
Uses the library GO.db.

\code{x} may be a \code{data.frame}.
In such case, GO ids are expected in its row names.
}
\examples{
getGOnames (c("GO:0000018", "GO:0000038", "BAD_GO"))

}
\seealso{
\code{\link{propagateGO}}, \code{\link{goLeaves}},
\code{\link{splitOntologies}}, \code{\link{getKEGGnames}},
\code{\link{getOntology}}
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\keyword{GO}
\keyword{names}
\keyword{ontology}
