% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GO_KEGG_names.R
\name{getKEGGnames}
\alias{getKEGGnames}
\title{Get KEGG  names}
\usage{
getKEGGnames(x, ...)
}
\arguments{
\item{x}{a character vector of KEGG ids.}

\item{verbose}{verbose.}
}
\value{
A character vector with the corresponding KEGG names.
}
\description{
Finds the KEGG name form KEGG id.
}
\details{
Uses the library KEGG.db.

\code{x} may be a \code{data.frame}.
In such case, GO ids are expected in its row names.
}
\examples{
getKEGGnames (c("00010", "00020", "BAD_KEGG"))

}
\seealso{
\code{\link{getGOnames}}
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\keyword{KEGG}
\keyword{names}
