% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mdPat}
\alias{mdPat}
\title{Multi-Dimensional Gene Set Analysis Pattern Classification.}
\usage{
mdPat(gsaout, cutoff = 0.05, pvalue = "padj")
}
\arguments{
\item{gsaout}{data.frame; output from mdGsa.}

\item{cutoff}{p-value cutoff for considering significant a Gene Set.}

\item{pvalue}{p-value column to be used. Default is named "padj" as in
mdGsa output.}
}
\value{
A character vector indicating the pattern associated to each
Gene Set.
}
\description{
Classifies significant patterns form a Multi-Variate Gene Set Analysis.
}
\details{
Sign of the three 'lor' and p-values are used to classify functional blocks.
The classification is done in the two dimensional space previously analyzed
by mdGsa.

All possible functional block classifications in the bi-dimensional
gene set analysis are:
\itemize{
  \item q1i: block displaced toward quadrant \bold{1} (0 < X & 0 < Y) with interaction.
  \item q2i: block displaced toward quadrant \bold{2} (0 > X & 0 < Y) with interaction.
  \item q3i: block displaced toward quadrant \bold{3} (0 > X & 0 > Y) with interaction.
  \item q4i: block displaced toward quadrant \bold{4} (0 < X & 0 > Y) with interaction.
  \item q1f: block displaced toward quadrant \bold{1}, no interaction.
  \item q2f: block displaced toward quadrant \bold{2}, no interaction.
  \item q3f: block displaced toward quadrant \bold{3}, no interaction.
  \item q4f: block displaced toward quadrant \bold{4}, no interaction.
  \item xh: block shifted to \bold{positive X} values.
  \item xl: block shifted to \bold{negative X} values.
  \item yh: block shifted to \bold{positive Y} values.
  \item yl: block shifted to \bold{negative Y} values.
  \item b13: bimodal block. Half of the genes displaced towards quadrant \bold{1} and the other half towards quadrant \bold{3}.
  \item b24: bimodal block. Half of the genes displaced towards quadrant \bold{2} and the other half towards quadrant \bold{4}.
  \item NS: \bold{non significant} block.
}
}
\examples{
N <- c (10, 20, 30, 40)
lor.X <- c (1.45, -0.32, 1.89, -1.66)
lor.Y <- c (2.36, -1.86, 0.43, -2.01)
lor.I <- c (0.89, -0.12, 0.24,  3.55)
pval.X <- c (0.001, 0.002, 0.003, 0.06)
pval.Y <- c (0.002, 0.003, 0.06,  0.07)
pval.I <- c (0.003, 0.02,  0.05,  0.08)
padj.X <- p.adjust (pval.X, "BY")
padj.Y <- p.adjust (pval.Y, "BY")
padj.I <- p.adjust (pval.I, "BY")

mdGsa.res <- as.data.frame (cbind (N,
                                   lor.X, lor.Y, lor.I,
                                   pval.X, pval.Y, pval.I,
                                   padj.X, padj.Y, padj.I))
mdGsa.res

mdGsa.res[,"pat"] <- mdPat (mdGsa.res)
mdGsa.res
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\references{
Montaner et al. (2010)
"Multidimensional Gene Set Analysis of Genomic Data."
PLoS ONE.
}
\seealso{
\code{\link{mdGsa}}, \code{\link{uvPat}}
}
\keyword{GSA}
\keyword{multidimensional}
\keyword{multivariate}
\keyword{pattern}

