% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_format.R
\name{revList}
\alias{revList}
\title{Revert an annotation list.}
\usage{
revList(lis, tag = "listPos")
}
\arguments{
\item{lis}{annotation list.}

\item{tag}{substitutes missing list names if any.}
}
\value{
An inverted list.
}
\description{
Inverts a list: names to elements / elements to names
}
\examples{
lis <- list (Block1 = c("gen1", "gen2"), Block2 = c("gen1", "gen3"))
revList (lis)

}
\seealso{
\code{\link{annotMat2list}}, \code{\link{revList}}
annotMat2list, annotList2mat
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\keyword{annotation}
\keyword{list}
\keyword{revert}
