% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uvSignif}
\alias{uvSignif}
\title{Filter significant terms in the univariate gene set analysis.}
\usage{
uvSignif(gsaout, cutoff = 0.05, pvalue = "padj", statistic = "lor",
  verbose = TRUE, sort = TRUE)
}
\arguments{
\item{gsaout}{data.frame; output from uvGsa.}

\item{cutoff}{p-value cutoff for considering significant a Gene Set.}

\item{pvalue}{p-value column to be used.
Default is named "padj" as in uvGsa output.}

\item{statistic}{name of the column containing the log odds ratio
from the uvGsa analysis.}

\item{verbose}{verbose}

\item{sort}{if TRUE the output data.frame is ordered according to
significance.}
}
\value{
The input data.frame but keeping just the
'significant' functional blocks.
}
\description{
Filters the rows in the data.frame returned by \code{\link{uvGsa}}
so that only the
enriched blocks are kept.
}
\details{
Works as \code{\link{goLeaves}} but without removing redundant terms.
}
\examples{
uvGsa.res <- as.data.frame (list (N    = c (10, 20, 30, 40),
                                  lor  = c (1.45, -0.32, 1.89, -1.66),
                                  pval = c (0.001, 0.002, 0.05, 0.06)))
uvGsa.res[,"padj"] <- p.adjust (uvGsa.res$pval, "BY")
uvGsa.res
uvSignif (uvGsa.res)

\dontrun{
res <- uvGsa (rindex, annotList)
uvSignif (res)
}
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\seealso{
\code{\link{uvGsa}}, \code{\link{uvPat}},
\code{\link{propagateGO}}, \code{\link{pval2index}}, \code{\link{goLeaves}}
}
\keyword{enriched}
\keyword{significant}
\keyword{terms}

